+++ src/Views/Traits/Helpers/ColumnHelpers.php
@@ -93,7 +93,7 @@ trait ColumnHelpers
     }
 
     // TODO: Test
-    public function renderContents(Model $row): mixed
+    public function renderContents(Model $row): null|string|\BackedEnum|HtmlString|DataTableConfigurationException|\Illuminate\Contracts\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View
     {
         if ($this->shouldCollapseOnMobile() && $this->shouldCollapseOnTablet()) {
             throw new DataTableConfigurationException('You should only specify a columns should collapse on mobile OR tablet, not both.');
@@ -102,8 +102,8 @@ trait ColumnHelpers
         return $this->getContents($row);
     }
 
-    // TODO: Improve TypeHinting & Tests
-    public function getContents(Model $row): mixed
+    // TODO: Test
+    public function getContents(Model $row): null|string|\BackedEnum|HtmlString|DataTableConfigurationException|\Illuminate\Contracts\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View
     {
         if ($this->isLabel()) {
             $value = call_user_func($this->getLabelCallback(), $row, $this);