+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 All notable changes to `laravel-livewire-tables` will be documented in this file
 
 ## UNRELEASED
-- Fix Return Type hinting to allow Enum columns
+- Fix Return Type hinting for Column Rendering to allow Enum columns
 
 ## [3.0.0-beta.4] - 2023-10-17
 - Introduction of Loading Placeholder